游戏开发完成

+++ app/src/main/java/com/civitasv/page/game/OnScoreChangeListener.java
@@ -2,5 +2,5 @@ package com.civitasv.page.game;
 
 
 public interface OnScoreChangeListener {
-    void onChange(int score);
+    void onChange(long score);
 }
