+++ app/src/main/java/com/civitasv/page/game/OnStateChangeListener.java
@@ -1,6 +1,6 @@
 package com.civitasv.page.game;
 
 public interface OnStateChangeListener {
-    void onWon(int score);
-    void onFail(int score);
+    void onWon(long score);
+    void onFail(long score);
 }
